Updated UI to scale with Screen size. Updated AI Prompts
Moved location of AI Prompt text object, adjusted font size. Made improvements with the interactions between slots, roulette and blackjack with the ApiManager, e.g. prompts are now made if the player loses in blackjack and prompts are not made if the player is doing AutoSpins on the slot machine to avoid spam. Added a setting to allow the player to decide if they want to have prompts enabled.